linting and renaming
