+++ Keas.Mvc/ClientApp/components/Access/AccessDetails.tsx
@@ -1,10 +1,10 @@
-﻿import * as React from 'react';
+import * as React from 'react';
+import { RouteComponentProps } from 'react-router';
 import { toast } from 'react-toastify';
 import { Button } from 'reactstrap';
 import { Context } from '../../Context';
 import { IAccess } from '../../Types';
 import AccessEditValues from './AccessEditValues';
-import { RouteComponentProps } from 'react-router';
 
 interface IProps {
   goBack: () => void;
